requirements.txt
# Core dependencies
accelerate==0.31.0
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.23.4
idna==3.4
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
pillow==10.2.0
psutil==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0   # upgraded to avoid conflicts
safetensors==0.4.3
six==1.16.0
sympy==1.12
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2     # updated, compatible with requests>=2.31.0

# PyTorch (MPS backend for Mac; avoids CUDA-only packages)
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
# ⚠️ Do NOT install triton (not supported on Mac M1)

# Optional for Hugging Face datasets
datasets==2.19.1
